+++ R/extract_ts.R
@@ -26,7 +26,7 @@
 #' data(selection_files)
 #' 
 #' #save 'Raven' selection tables in the temporary directory 
-#' writeLines(selection_files[[5]], con = names(selection_files)[5])
+#' writeLines(selection_files[[5]], con = file.path(tempdir(), names(selection_files)[5]))
 #'
 #' \donttest{
 #' # import data to R