+++ R/imp_corr_mat.R
@@ -19,18 +19,19 @@
 #' \dontrun{ 
 #' # Load data
 #' library(warbleR)
-#' data(list = c("Phae.long1", "Phae.long2", "Phae.long3", "Phae.long4", "selec.table"))
+#' data(list = c("Phae.long1", "Phae.long2", "Phae.long3", "Phae.long4", "lbh_selec_table"))
 #' 
-#' writeWave(Phae.long1, "Phae.long1.wav", extensible = FALSE) #save sound files 
-#' writeWave(Phae.long2, "Phae.long2.wav", extensible = FALSE)
-#' writeWave(Phae.long3, "Phae.long3.wav", extensible = FALSE)
-#' writeWave(Phae.long4, "Phae.long4.wav", extensible = FALSE)
+#' writeWave(Phae.long1, file.path(tempdir(), "Phae.long1.wav"), extensible = FALSE) #save sound files 
+#' writeWave(Phae.long2, file.path(tempdir(), "Phae.long2.wav"), extensible = FALSE)
+#' writeWave(Phae.long3, file.path(tempdir(), "Phae.long3.wav"), extensible = FALSE)
+#' writeWave(Phae.long4, file.path(tempdir(), "Phae.long4.wav"), extensible = FALSE)
 #' 
 #' #create new folder to put cuts
-#' dir.create("cuts")
+#' dir.create(file.path(tempdir(), "cuts"))
 #' 
 #' # cut files
-#' cut_sels(X = selec.table, mar = 0.05, path = tempdir(), dest.path = file.path(tempdir(), "cuts"))
+#' cut_sels(X = lbh_selec_table, mar = 0.05, path = tempdir(),
+#'  dest.path = file.path(tempdir(), "cuts"))
 #' 
 #' #Now run 'Raven' batch correlator un the cuts and save the output in the same folder
 #' 
@@ -48,19 +49,15 @@
 #' }
 #' 
 #' @author Marcelo Araya-Salas (\email{marceloa27@@gmail.com})
-#last modification on nov-7-2017
+#last modification on sep-2-2019
 imp_corr_mat <- function(file, path = NULL){
   
-  # reset working directory 
-  wd <- getwd()
-  on.exit(setwd(wd))
-  
   #check path to working directory
-  if (is.null(path)) path <- getwd() else {if (!dir.exists(path)) stop("'path' provided does not exist") else
-    setwd(path)
-  }  
+  if (is.null(path)) path <- getwd() else 
+    if (!dir.exists(path)) 
+      stop("'path' provided does not exist") 
   
-  lns <- readLines(file)
+  lns <- readLines(file.path(path, file))
   lbks <- which(lns == "")
   
   mat1 <- do.call(rbind, strsplit(lns[lbks[1]:lbks[2]], split = "\t"))