+++ R/relabel_colms.R
@@ -26,11 +26,11 @@
 #' data(selection_files)
 #' 
 #' #save 'Raven' selection tables in the temporary directory 
-#' writeLines(selection_files[[5]], con = names(selection_files)[5])
+#' writeLines(selection_files[[5]], con = file.path(tempdir(), names(selection_files)[5]))
 #'
 #' \donttest{ 
 #' #'# import data to R
-#'rvn.dat <- imp_raven(all.data = TRUE) 
+#'rvn.dat <- imp_raven(all.data = TRUE, path = tempdir()) 
 #'
 #' names(rvn.dat)
 #'